+++ www/go/base/db/ActiveRecord.php
@@ -4738,6 +4738,7 @@ abstract class ActiveRecord extends \GO\Base\Model{
 					} catch (\Exception $e) {
 						\go\core\ErrorHandler::logException($e);
 						echo "E";
+						$start++;
 					}
 				}
 				echo "\n";