+++ www/go/core/views/extjs3/links/LinksDetailPanel.js
@@ -127,7 +127,7 @@ go.links.DetailPanel = Ext.extend(Ext.Panel, {
 });
 
 go.links.getDetailPanels = function() {
-
+	
 	var panels = [];
 	go.links.Links.getAll().forEach(function (e) {		
 		panels.push(new go.links.DetailPanel({