+++ www/modules/calendar/EventDialog.js
@@ -547,13 +547,13 @@ Ext.extend(GO.calendar.EventDialog, Ext.util.Observable, {
 				if(action.result.feedback){
 					Ext.MessageBox.alert(t("Error"), action.result.feedback);
 				}else	if (hide) {
-					this.win.close();					
+					this.win[this.win.closeAction]();	
 				}
 
 				if (config && config.callback) {
 					config.callback.call(this, this, true);
 				}
-				
+
 				if(this.win.closeAction != "close") {					
 					this.participantsPanel.store.loadData({results:action.result.participants});
 				}